requirements.txt
wandb==0.16.3
numpy==1.26.4
einops==0.7.0
torch==2.2.1
torchvision==0.17.1
earthengine-api==0.1.391
rioxarray==0.15.1
rasterio==1.3.10
tqdm==4.66.2
geopandas==0.14.3
fiona==1.9.6
geobench==1.0.0
h5py==3.10.0
scikit-learn==1.4.1.post1
types-requests
# for evals
fast_pytorch_kmeans
satlaspretrain-models
timm
breizhcrops
